Add bottom navigation bar, improve flow and UI
Add bottom navigation bar with customizable menu. Implement relevant fragment java and xml files as well as the switch in the activity.

Improve the flow and UI such that buttons appear and disappear based on program state (might need to further improve this, this way is a bit unintuitive). Improve UI in general.

Implement a basic User class to represent users.

Change color theme.